functional_tests: use `size_t` for size variable
Changed the data type of `size` to `size_t` in all functional tests to ensure
consistency with rocSHMEM APIs.

+++ tests/functional_tests/amo_extended_tester.cpp
@@ -55,7 +55,7 @@ AMOExtendedTester<T>::~AMOExtendedTester() {
 }
 
 template <typename T>
-void AMOExtendedTester<T>::resetBuffers(uint64_t size) {
+void AMOExtendedTester<T>::resetBuffers(size_t size) {
   memset(_r_buf, 0, args.max_msg_size);
   memset(_ret_val, 0, args.max_msg_size * args.num_wgs);
   memset(_s_buf, 0, args.max_msg_size * args.num_wgs);
@@ -63,7 +63,7 @@ void AMOExtendedTester<T>::resetBuffers(uint64_t size) {
 
 template <typename T>
 void AMOExtendedTester<T>::launchKernel(dim3 gridsize, dim3 blocksize, int loop,
-                                        uint64_t size) {
+                                        size_t size) {
   size_t shared_bytes = 0;
 
   hipLaunchKernelGGL(AMOExtendedTest, gridsize, blocksize, shared_bytes, stream,
@@ -82,7 +82,7 @@ void AMOExtendedTester<T>::launchKernel(dim3 gridsize, dim3 blocksize, int loop,
 #endif
 
 template <typename T>
-void AMOExtendedTester<T>::verifyResults(uint64_t size) {
+void AMOExtendedTester<T>::verifyResults(size_t size) {
   T ret;
   if (args.myid == 0) {
     T expected_val = 0;
