+++ tests/functional_tests/team_alltoall_tester.cpp
@@ -153,7 +153,7 @@ void TeamAlltoallTester<T1>::preLaunchKernel() {
 
 template <typename T1>
 void TeamAlltoallTester<T1>::launchKernel(dim3 gridSize, dim3 blockSize,
-                                          int loop, uint64_t size) {
+                                          int loop, size_t size) {
   size_t shared_bytes = 0;
 
   int num_elems = size / sizeof(T1);
@@ -175,7 +175,7 @@ void TeamAlltoallTester<T1>::postLaunchKernel() {
 }
 
 template <typename T1>
-void TeamAlltoallTester<T1>::resetBuffers(uint64_t size) {
+void TeamAlltoallTester<T1>::resetBuffers(size_t size) {
 
   int num_elems = size / sizeof(T1);
   int buff_size = num_elems * sizeof(T1) * args.num_wgs * n_pes;
@@ -204,7 +204,7 @@ void TeamAlltoallTester<T1>::resetBuffers(uint64_t size) {
 }
 
 template <typename T1>
-void TeamAlltoallTester<T1>::verifyResults(uint64_t size) {
+void TeamAlltoallTester<T1>::verifyResults(size_t size) {
   int num_elems = size / sizeof(T1);
   int idx = 0;
