+++ tests/functional_tests/team_ctx_infra_tester.cpp
@@ -113,7 +113,7 @@ TeamCtxInfraTester::TeamCtxInfraTester(TesterArguments args) : Tester(args) {}
 
 TeamCtxInfraTester::~TeamCtxInfraTester() {}
 
-void TeamCtxInfraTester::resetBuffers(uint64_t size) {}
+void TeamCtxInfraTester::resetBuffers(size_t size) {}
 
 void TeamCtxInfraTester::preLaunchKernel() {
   int n_pes = rocshmem_team_n_pes(ROCSHMEM_TEAM_WORLD);
@@ -149,7 +149,7 @@ void TeamCtxInfraTester::preLaunchKernel() {
 }
 
 void TeamCtxInfraTester::launchKernel(dim3 gridSize, dim3 blockSize, int loop,
-                                      uint64_t size) {
+                                      size_t size) {
   size_t shared_bytes = 0;
 
   /* Copy array of teams to device */
@@ -170,4 +170,4 @@ void TeamCtxInfraTester::postLaunchKernel() {
   }
 }
 
-void TeamCtxInfraTester::verifyResults(uint64_t size) {}
+void TeamCtxInfraTester::verifyResults(size_t size) {}
